+++ mods/Farming/mod/data/herb_data.lua
@@ -1,78 +1,78 @@
 return {
   ["guam"] = {
     level = 1,
-    tint = {0, 161, 0}
+    tint = {r=0, g=161, b=0}
   },
   ["tarromin"] = {
     level = 5,
-    tint = {0, 214, 0}
+    tint = {r=0, g=214, b=0}
   },
   ["marrentill"] = {
     level = 9,
-    tint = {0, 222, 0}
+    tint = {r=0, g=222, b=0}
   },
   ["harralander"] = {
     level = 20,
-    tint = {31, 167, 0}
+    tint = {r=31, g=167, b=0}
   },
   ["ranarr"] = {
     level = 25,
-    tint = {7, 147, 0}
+    tint = {r=7, g=147, b=0}
   },
   ["toadflax"] = {
     level = 30,
-    tint = {1, 103, 0}
+    tint = {r=1, g=103, b=0}
   },
   ["spirit-weed"] = {
     level = 35,
-    tint = {129, 204, 167}
+    tint = {r=129, g=204, b=167}
   },
   ["irit"] = {
     level = 40,
-    tint = {4, 167, 0}
+    tint = {r=4, g=167, b=0}
   },
   ["wergali"] = {
     level = 41,
-    tint = {228, 7, 0}
+    tint = {r=228, g=7, b=0}
   },
   ["avantoe"] = {
     level = 48,
-    tint = {0, 131, 0}
+    tint = {r=0, g=131, b=0}
   },
   ["kwuarm"] = {
     level = 54,
-    tint = {61, 114, 0}
+    tint = {r=61, g=114, b=0}
   },
   ["bloodweed"] = {
     level = 57,
-    tint = {162, 22, 0}
+    tint = {r=162, g=22, b=0}
   },
   ["snapdragon"] = {
     level = 59,
-    tint = {53, 119, 0}
+    tint = {r=53, g=119, b=0}
   },
   ["cadantine"] = {
     level = 65,
-    tint = {45, 107, 0}
+    tint = {r=45, g=107, b=0}
   },
   ["lantadyme"] = {
     level = 67,
-    tint = {0, 153, 55}
+    tint = {r=0, g=153, b=55}
   },
   ["dwarf-weed"] = {
     level = 70,
-    tint = {0, 141, 0}
+    tint = {r=0, g=141, b=0}
   },
   ["torstol"] = {
     level = 75,
-    tint = {0, 187, 0}
+    tint = {r=0, g=187, b=0}
   },
   ["arbuck"] = {
     level = 77,
-    tint = {235, 88, 0}
+    tint = {r=235, g=88, b=0}
   },
   ["fellstalk"] = {
     level = 91,
-    tint = {17, 82, 7}
+    tint = {r=17, g=82, b=7}
   }
 }